uint ALL the things!

+++ Hasse/Groups/Heavy/Permutation/SymmetricElement.cs
@@ -3,21 +3,21 @@ using System.Text;
 
 namespace Hasse.Groups.Heavy.Permutation{
 	public class SymmetricElement : GroupElement<SymmetricElement>{
-		private int[] Value{get;set;}
+		private uint[] Value{get;set;}
 
-		public SymmetricElement(int[] value){
+		public SymmetricElement(uint[] value){
 			Value = value;
 		}
 
 		public override SymmetricElement Multiply(SymmetricElement other){
-			int[] res = new int[Value.Length];
-			for(int i = 0; i < Value.Length; i++)
+			uint[] res = new uint[Value.Length];
+			for(uint i = 0; i < Value.Length; i++)
 				res[i] = other.Value[Value[i]];
 			return new SymmetricElement(res);
 		}
 
 		public override bool Equals(SymmetricElement pobj){
-			for(int i = 0; i < Value.Length; i++)
+			for(uint i = 0; i < Value.Length; i++)
 				if(Value[i] != pobj.Value[i])
 					return false;
 			return true;
@@ -27,7 +27,7 @@ namespace Hasse.Groups.Heavy.Permutation{
 			int toret = 0;
 			for(int i = 0; i < Value.Length; i++){
 				toret *= Value.Length;
-				toret += Value[i];
+				toret += (int)Value[i];
 			}
 			return toret;
 		}
@@ -35,11 +35,11 @@ namespace Hasse.Groups.Heavy.Permutation{
 		public override string ToString(){
 			bool[] done = new bool[Value.Length];
 			StringBuilder sb = new StringBuilder();
-			for(int i = 0; i < Value.Length; i++)
+			for(uint i = 0; i < Value.Length; i++)
 				if(!done[i] && Value[i] != i){
 					sb.Append('(');
 					sb.Append(i + 1);
-					for(int curr = Value[i]; curr != i; curr=Value[curr]){
+					for(uint curr = Value[i]; curr != i; curr=Value[curr]){
 						done[curr] = true;
 						sb.Append(' ');
 						sb.Append(curr + 1);
